+++ services/py-api/src/database/repository/feature_switch_repository.py
@@ -7,7 +7,7 @@ from src.database.db_manager import MongoDatabaseManager
 from src.database.model.feature_switch_model import FeatureSwitch
 from structlog.stdlib import get_logger
 from src.database.repository.base_repository import CRUDRepository
-from src.server.exception import FeatureSwitchNotFoundError
+from src.exception import FeatureSwitchNotFoundError
 
 LOG = get_logger()
