+++ services/py-api/src/server/server_config.py
@@ -7,7 +7,7 @@ from uvicorn import run
 
 from src.database.db_manager import ping_db
 from src.environment import DOMAIN, PORT
-from src.server.logger.logger_factory import get_uvicorn_logger
+from src.logger.logger_factory import get_uvicorn_logger
 from src.utils import singleton
 
 LOG = get_logger()
