+++ services/py-api/src/service/feature_switch_service.py
@@ -3,7 +3,7 @@ from typing import List
 from result import Result
 from src.database.model.feature_switch_model import FeatureSwitch
 from src.database.repository.feature_switch_repository import FeatureSwitchRepository
-from src.server.exception import FeatureSwitchNotFoundError
+from src.exception import FeatureSwitchNotFoundError
 
 
 class FeatureSwitchService:
