+++ services/py-api/src/service/hackathon/hackathon_mail_service.py
@@ -28,7 +28,7 @@ class HackathonMailService:
         verification_link: str,
         background_tasks: BackgroundTasks,
         team_name: str | None = None,
-    ) -> Err[ValueError]:
+    ) -> Err[ValueError] | None:
         """
         Sends a verification email to admin and random participants via FastAPI BackgroundTasks
 
@@ -68,13 +68,15 @@ class HackathonMailService:
             content_type="html",
         )
 
+        return None
+
     def send_participant_successful_registration_email(
         self,
         participant: Participant,
         background_tasks: BackgroundTasks,
         invite_link: str | None = None,
         team_name: str | None = None,
-    ) -> Err[ValueError]:
+    ) -> Err[ValueError] | None:
         """
         Sends an email confirming the successful registration of a participant.
 
@@ -114,6 +116,8 @@ class HackathonMailService:
             body_content=body_html,
             content_type="html",
         )
+
+        return None
 
 
 def hackathon_mail_service_provider(client: MailClient) -> HackathonMailService:
